Clean up
* removed unnecessary casts
* removed effectless assignments x = x
* removed unnecessary null checks

+++ system/src/de/uka/ilkd/key/logic/PosInOccurrence.java
@@ -142,7 +142,7 @@ public class PosInOccurrence {
     }
 
     public int hashCode () {
-    	return (int) hashCode;
+    	return hashCode;
     }  
 
     /**
